+++ packages/expo-audio/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### 🎉 New features
 
+- Support creating an `AudioPlayer` instance without using the `useAudioPlayer` hook.
+
 ### 🐛 Bug fixes
 
 - Update docs and API so all units of time are returned as seconds, not milliseconds. ([#33320](https://github.com/expo/expo/pull/33320) by [@alanjhughes](https://github.com/alanjhughes))